netplantest: move under dbusutil
